Log client transactions and give context for errors (#426)
Improved contract transaction logging and error. 

It is implemented by wrapping a `Contract`'s functions with delegator functions. The properties of the contract object are read-only -> it is not possible to modify the existing function definitions. (Also the functions are implemented as anonymous functions, and therefore it is not possible to get the function name just by querying  `function.name`).

### New features

If a transaction method execution (call or send) throws an `Exception`, the message contains the method name in the error message. E.g.:
```
Error: Error in contract call "streamRegistry.createStream"
  ...
  reason: Error: could not detect network (event="noNetwork", code=NETWORK_ERROR, version=providers/5.5.2)
  ...
```

Transaction log contains the name of the contract method. E.g.:
```
transaction submitted { method=streamStorageRegistry.addStorageNode, ... }
transaction confirmed { method=streamStorageRegistry.addStorageNode, ... }
```

The Ethereum address is printed as a human-readable name if known. E.g.: 
```
transaction submitted { ..., to=DOCKER_DEV_STREAM_STORAGE_REGISTRY, ... }
```

There are some human-readable addresses configured in `network/NameDirectory.ts`. It helps us to identify the services in Streamr docker dev environment. In other environments (and if the human-readable name is not known) the address will be printed as shortened Ethereum address. E.g.:
```
transaction submitted { ... to=0x123456, ... }
```

All transaction method executions are logged. E.g.:
```
execute streamRegistry.getStreamMetadata
```


### Future improvements
- We could use these logging + error handling improvements in all contract handling? (e.g. DataUnion contract and tracker registry)
- We could add all docker-dev addresses to `NameDirectory` if we find the human-readable names useful. Maybe that makes sense when the list of addresses are easily available in some always up-to-date JSON.