Client fake test environment (#432)
Added fake versions of BrubeckNode and blockchain registry facades. We can use these classes in tests to run the test in a fake environment. In the fake environment all stream-related smart contract calls (StreamRegistry and StorageNodeRegistry) and messaging between nodes are emulated. 

Using a fake environment makes some tests are significantly faster. In this PR we applied the fake environment for two slow tests: 
- Encryption.test.ts: speed-up 420s -> 57s
- GroupKeyPersistence.test.ts: speed-up 225s -> 9s

In those two tests the functionality under the test is not related to smart contracts or the messaging between the nodes (the tests send messages, but there is nothing special about the messaging -> no need to test that aspect here).

We can make separate PRs for applying the fake environment for other tests. We should not apply the fake environment for all integration tests (e.g. if the main focus of a test is messaging and/or smart contracts), but maybe it is applicable for some of the tests.

### Implementation

We apply the fake environment to a test by creating StreamrClient instances via custom ClientFactory. We get the factory by calling createClientFactory (fakeEnvironment/clientFactory.ts). It uses dependency injection to inject some objects (mostly singletons). We can create multiple clients from each clientFactory: all instances from a factory share the same fake environment. For each test case we create a new factory and therefore get a clean fake environment that is independent of other test cases.

The tests don't need significant modification when we convert a "real" test to a fake test. Usually we can just change the way we make the client instances: e.g. call clientFactory.createClient() instead of getCreateClient().createClient() (which is helper method in test-utils/utils.ts). In the fake environment we can authenticate a client with any private key, there is no need to use fetchPrivateKeyWithGas().

There are four direct replacements for StreamrClient dependencies:
- FakeBrubeckNode emulates BrubeckNode
- FakeStreamRegistry emulates StreamRegistry
- FakeStorageNodeRegistry emulates StorageNodeRegistry
- FakeRest emulates Rest

These fake classes implement a subset of the functionality of "real" classes. For some methods the methods may throw a "not implemented" exception if we have not yet implemented the functionality in the fake class. We need implementations only for those methods which are actually used in the test.

Each class implements the interface of the corresponding class and therefore ensures that the method definitions are kept in sync between the two implementations.

There are also two helper classes:
- FakeStorageNode which emulates the storage node functionality
- ActiveNodes which keeps track of FakeBrubeckNode instances (contains all nodes created for a fake environment)

FakeStorageNodeRegistry creates one storage node with docker-dev environment node address (0xde1112…). There is currently no implementation for createOrUpdateNodeInStorageNodeRegistry, but we can add that later if custom storage nodes are needed.

### Small changes for the core StreamClient internals

Small refactoring was needed to allow the faking of some classes:
- Resends.fetchStream moved to Rest
- BrubeckNode.start returns Promise<void> (instead of a promise which leaks a NetworkNode)

Also the return type of BrubeckNod.getNode() changed from NetworkNode to a new interface called NetworkNodeStub. It is the same object as previously, but by annotating the instance with the interface we limit the public API of the client. Most of NetworkNode methods/fields are internal implementation and there is no need to expose those to users of the client. Alternatively we could mark some methods/fields of NetworkNode as @internal./private Or maybe we could provide the functionality of NetworkNode methods as StreamrClient methods.

### Other small bugfixes/refactoring

These are not strictly related to the fake implementation/refactoring:
- FailedToPublishError error message tweaked so that it doesn't try to make a JSON from a full Stream object (if it is provided as streamDefinition parameter). A full object may contain circular references and therefore stringifying it can fail.
- Changed field visibility in StreamRegistry and StorageNodeRegistry, removed redundant read-only contract in StorageNodeRegistry
- Fixed return type of StreamEndpoints.getStreamLast
- test/utils.ts moved to test/test-utils/utils.ts
- added Multimap helper class to test utils

### Future improvements

Do we want to as a separate test/end-to-end and move some tests there? In that setup there would be three tests types:
- Unit tests to test one class directly by calling the class methods.
- Integration tests to test some larger functionality by executing methods via StreamrClient (usually interaction between nodes and/or smart contracts). Usually testing multiple edge cases related to that functionality.
- End-to-end tests to test that all Streamr components (especially calls to blockchain registries and Core-API etc.) work well together and we set up the usages of our external services correctly. Usually focusing on happy paths, but also full coverage for the smart contract usage.