[NET-595]: Migrate Testnet3 broker config (#446)
Automatically migrate a Testnet3 config to the new format used in Brubeck mainnet.

The migration drops many properties, either because the new value is already the default in the config schema or because the value was Testnet3-specific. See migration.ts:convertTestnet3ToV1() method for details.

### Schema

There is a new $schema property in the config file, which identifies the version of the config. For migrated configs, and new configs created with Config Wizard, the reference is http://schema.streamr.com/config-1.schema.json

That value is JSON-schema compatible URL. In the migration we use the filename part of the URL to investigate
the config file version. We don't fetch the content of the schema file at all. The benefit of using schema  instead of just a version number is that it provides more information to any external tools, e.g. IDEs. If we upload the schema file to some public server, an IDE can provide validation and autocompletion if a user edits a configuration file manually.

We aren't expected to change the schema format very often. Therefore the uploading of the schema file could be just a manual task. It is also not a critical task: the migration works fine even we aren't able to upload the file for some reason. 

### File location

User can specify the config file location as a command line parameter:
- If it specified, we create the migrated file to that location and copy a backup file to the same directory.
- If it is not specified, we check the current default location ~/.streamr/config/default.json and the legacy default location ~/.streamr/broker-config.json. If the migration has not been done yet, the file is very likely found from the legacy location, and there is no file in the current default location. In that situation we create the migrated file to the current default location. We also "archive" the legacy file by renaming it. The new broker-config.json.backup name communicates that the legacy file is no longer used. (The renaming is the last step of the migration, and therefore we don't do that if an earlier step fails.)

### Testing

There is an extensive unit test for the migration logic, but no tests for the file handling (reading the config file, creating a backup, writing a new file). We could add a test if we implement some kind of file system emulation (e.g. by using [mock-fs](https://www.npmjs.com/package/mock-fs). Maybe not worth doing as this it is relatively easy to test this manually and we don't expect that the implementation is modified very often?