client: validate searchStreams params (#461)
If there is no search term or permission filter, `client.searchStream` throws an `Error`. 

It could return all streams from the registry, but maybe it is better to an error. In an earlier PR we removed the `client.getAllStreams()` method -> it doesn't make sense that we'd support the "search all streams" functionality via this method.