test: maybe fix flaky test for test-utils.wait() (#469)
**Eric's writing:**
I'm gonna go on a gut feeling here and replace this to work with `Date.now()` instead of `performance.now()`. My recollection is that it was like that in the past and was not flaky. I have no idea why taking the time in a more accurate manner causes this, but yeah, if this test stays flaky let's re-look into it.