refactor: remove redundant dependency injections (#496)
Remove injected objects which were not used by the client. 

One of the fields is `client.options`, which contains e.g. the private key. The removal also enhances security. If we pass a `StreamrClient` instance to some external component, that component shouldn't be able to read the private key.

Refactor some test which used `client.options` directly. In `MemoryLeak.test.ts` there was a leak detection exception for `client.config.network`, but that is maybe not needed anymore? The `client.config.network` object contains just the configuration options, not e.g. a `NetworkNode` instance.