client [NET-606]: Wait for The Graph to sync when necessary  (#502)
If we write to `StreamRegistry` or `StorageNodeRegistry` via client methods, all subsequent reads from The Graph provide up-to-date responses automatically. 

Before this PR there was an internal method `streamExistsOnTheGraph` to do the poll the stream creation, but not similar methods for permissions etc. That kind of manual polling is no longer needed as `SynchronizedGraphQLClient` polls the state of The Graph.

Typical scenario for the synchronization:
1.  First we write info to the blockchain, e.g. call `grantPermissions()`:  
    - extract the block number from the transaction receipt (notified by `ObservableContract`'s event emitter)
    - update  `SynchronizedGraphQLClient`'s `requiredBlockNumber`
2. After that, we query the related data, e.g. call `getPermissions()`:
    - start polling the indexing state from The Graph
    - wait until  The Graph metadata that it has indexed the required block (`blockNumber`>=`requiredBlockNumber`)
    - execute the query and return the data to the caller

If we have not written anything to the blockchain via client methods, The Graph polling is not needed and `SynchronizedGraphQLClient` executes the queries immediately.

We manage the waiting by using `Gate` instances. Each queried `blockNumber` corresponds a gate, and the poller opens a gate when it sees that The Graph has indexed the waited block. When we add gates, we manually start the poller if needed. It automatically stops when there are no more work to do.

## Other

- Add `jest-extended` library (needed in `Permissions.test.ts`)
- Fix CI setup for `test-integration`

## Future improvements

-  If we need, we can later add functionality where users can query The Graph without waiting the sync. It is easy to add but would make the API more complex -> let's add it if/when there is a need for that.
- There is an implementation of `wait()` method in` SynchronizedGraphQLClient.ts`. It would be good to import that method a library (or alterntively move it to `utils.ts` and import from there). We have an implementation for the method in`streamr-test-utils`, but maybe we shouldn't include that library as a non-test dependency? The `streamr-test-utils` has dependency e.g. to  the `express` library, which is not needed in `StreamrClient` and can increase the bundle size unnecessarily.

## Checklist

- [x] Has passing tests that demonstrate this change works