client: Storage node assignment event listeners (#531)
Refactor storage node assignment event listeners to use EventEmitter style:
-  add a listener with `on('addToStorageNode')` / `on('removeFromStorageNode')` instead of `registerStorageEventListeners()`
- remove a listener with `off('addToStorageNode')` / `off('removeFromStorageNode')` instead of `unregisterStorageEventListeners()`

The `unRegisterStorageEventListeners()` removed all listeners. The `off('...', listener)` call always removes a specific listener.

Previously there was only one event, which had a field to indicate whether the event is an addition or a removal. Now it is possible to listen either or both of the events.

## Implementation

We get storage node events from the StorageNode contract by adding a listener to that contract. Adding a listener starts a poller which queries the blockchain state in 4 second intervals. 

To avoid unnecessary polling, we add a contract listener only when `StreamrClient` needs the event information. That is, when `client.on('addToStorageNode')` or `client.on('removeFromStorageNode)` is called.  

To support that kind of dynamic listener binding, there is an observable `StreamrClientEventEmitter` which emits an event whenever new listeners are added to a `StreamrClient` instance. The `initEventGateway` helper method is used to do the actual binding (binding the listener in `start` callback, and unbinding it in `stop` callback).

## Future improvements

- refactor `onResent` to use EventEmitter style
- automatically remove listeners when `destroy()` is called (e.g. by using `destroySignal.onDestroy()`)
- use generics in `EventEmitter3` instead of `strict-event-emitter-types` in client

## Checklist

- [x] Has passing tests that demonstrate this change works
- [x] Updated Changelog