fix(client): Publish encrypted messages (#574)
Removed a guard in `Encryption.ts` which caused message encryption to be skipped. The check was modified in https://github.com/streamr-dev/network-monorepo/commit/5d4e11e34a96761c7e7cdc776b559c23ec8f5e47, and it seems that the guard is no longer relevant as all private streams should be encrypted (i.e. stream.requireEncryptedData would be true, it that field still would be available). 

The encryption of public streams is checked separately in line 51:
```
const isPublic = await this.streamEndpoints.isPublic(streamId)
if (isPublic || this.isStopped) {
    return
}
```

## Mutability

When we decrypted message, we would mutate the `StreamMessage` so that its content went from encrypted to decrypted. If that message got sent to further nodes (which could happen if the the decryption process occurred before message propagation), it would go out to neighbors in cleartext. This would mean that upon receiving said message, signature validation would fail on neighbors. 

## Possible future improvements

- Any problems with `WeakSet` in SubscribePipeline class?
- Should we also clone `StreamMessage` before encrypting? It would be symmetrical to the decrypt case, and in a sense, less mutability in system.
- If we make `StreamMessage` to be immutable, the cloning can be removed

## Checklist

- [x] Has passing tests that demonstrate this change works
- [x] Updated Changelog