fix: WebRTC compatibility in Firefox (#593)
Improve WebRTC connection handling to be compatible with Firefox. 

We used `peerConnection.connectionState` to determine the state of a WebRTC connection, however this is not supported in some browsers, such as Firefox (See: https://developer.mozilla.org/en-US/docs/Web/API/RTCPeerConnection/connectionState#browser_compatibility and https://developer.mozilla.org/en-US/docs/Web/API/RTCPeerConnection/connectionstatechange_event). 

If it is not supported, it will have value undefined, which will trip up the logic to throw an error on 2nd access to the connection in https://github.com/streamr-dev/network-monorepo/blob/main/packages/network/src/connection/webrtc/WebRtcEndpoint.ts#L354. The 2nd access is typically made when a stream is encrypted as we deliver the group keys via the same connection.