broker [NET-839]:  Publisher plugins set msgChainId (#679)
Each publisher plugin sets `msgChainId` when it publishes messages via `StreamrClient`. For stateful plugins (`mqtt`, `websocket`), the `msgChainId` is connection-specific. For `http` plugin the `msgChain` is generated at the startup of the plugin.

### Implementation

`mqtt` plugin uses Aedes's `client.id` field as `msgChainId`. The value to that field is read from an incoming connection  packet. The spec required that there can't be two clients connected with same clientId simultanously. Users of MQTT client libraries can possibly set a custom value to that field, and in that situation the `client.id` wouldn't be globally unique. If a custom value is not used, the value is typically very unique:
- spec: "3.1.3.1 Client Identifier" spec in http://docs.oasis-open.org/mqtt/mqtt/v3.1.1/os/mqtt-v3.1.1-os.html 
- implementation is Aedes: https://github.com/moscajs/aedes/blob/e3e50ded602e7d69bdf8ba15ee07ce3527e004f1/lib/handlers/connect.js#L78
- e.g. `clientId` option settable in https://github.com/mqttjs/MQTT.js client library, but defaults to a unique identifier

### Future improvements

- `http` plugin could use cookies to emulate "connections", and assign connection-specific `msgChainId` values that way (if needed in the future)