refactor(protocol): KeyExchangeStream partition  (#708)
Changed `StreamIDUtils.formKeyExchangeStreamID` to return `StreamPartID` instead of `StreamPart`. 

The method is used by `StreamrClient`. Therefore when the client publishes/subscribes to a key exchange stream, it uses `0` partition explicitly . The functionality doesn't change as publish and subscribe currently use `0` partition by default if no partition is specified. But after this PR, key exchange operations don't need to rely on that default behavior.

Also extracted key exchange stream id utility methods to a separate `KeyExchangeStreamIDUtils` utils. As one of the methods doesn't operate on `StreamID` but `StreamPartID`, it wouldn't be optimal to leave key exchange methods to 
 be part `StreamIDUtils`. This change also simplifies the basic usage of `StreamIDUtils`.