refactor(client): Remove message container (#713)
Add new options for `MessageCreator` so that it can create messages with custom `messageType` or `encryptionType`.  `KeyExchangeStream` and `PublisherKeyExchangeStream` use those options to create group key messages. The options are marked as `@internal`, because we don't need to expose this functionality to end-users.

Earlier `GroupKeyMessages` were created via `StreamMessageContainer`. That functionality is now removed as it is no longer needed.