refactor(client): Simplify encrypt (#717)
Removed redundant call to `streamRegistryCached#getStream` as the method already knows streamId.  The reason for the original call is unclear. If it was there to check the existence of the stream, it is maybe no longer needed here as the existence is implicitly checked when we query from `StreamRegistry` that the publisher has the publish permission.

Removed duplicate `messageType` check. We already have a guard `(messageType !== MESSAGE_TYPES.MESSAGE)`.

Removed redundant call to Ethereum#canEncrypt. In practice that returned false if user wasn't authenticated. The encryption logic doesn't need e.g. user's ethereum address, and therefore we don't need to check that here. We already use address in `MessageCreator:62`, and therefore publishing without authentication fails as expected.

Removed obsolete `Ethereum#getSigner` method. It was not called by any component (e.g. `Signer#getSigningFunction` could maybe call it, but it doesn't). The method `getStreamRegistryChainSigner` was not removed as it is used by `StreamRegistry`.