feat(client): Remove `sequenceNumber` metadata option (#733)
Removed `sequenceNumber` from `MessageMetadata` type, which is used by `StreamrClient#publish`. 

`MessageCreator` doesn't read it and therefore it is ignored when we create a message.

As the field is not currently used by any component (and it is not part of documented public API), we can remove it. If we need the field in the future, we should implement support for it in `MessageCreator` .