refactor(client): Remove `onDestroy` cache clearing (#740)
When `client.destroy()` is called, it cleans up many resources. As it is not allowed to interact with the client after `destroy()` call, it is very likely that the `StreamrClient` will be garbage collected from memory very soon after the call (if the application doesn't keep a reference to it). Therefore it is not necessary need to explicitly clean up some of the memory resources which `StreamrClient` has allocated.

In this PR we removed cache clearing functionality which was triggered via `client.destroy()`.

Before this PR there were explicit cache clearing for these methods:
- `MessageCreator#getMsgChain`
- `StreamPartitioner#computeStreamPartition`
- `Validator#cachedVerify`
- `GroupKeyStoreFactory.getStore`

There was also a `StreamRegistryCached#clear` method, which was not called by any component (all callers used `StreamRegistryCached#clearStream`)

### Future improvements

- Each component could listen `destroySignal` independently, and clean up timers/promises it allocated. Currently it is typical that an external component calls a `stop()` method of a component to trigger the cleanup.
- We could rename all `stop()` methods to `destroy()`
- We could convert **test/integration/Partitions.test.ts** to unit test