refactor(client): Move chain configs from root (#752)
Moved chain address configs from root config interface to `EthereumConfig` interface. This change doesn't affect the config structure or format of `StrictStreamrClientConfig`, because that is the combination of both of the interfaces. But now dependencies are more explicit. E.g. `StreamRegistry` needs only a dependency to `EthereumConfig`, not to the root config.