refactor(client): Group key validation (#756)
Removed calls to `GroupKey#validate`. A `GroupKey` object is always in valid state, because the constructor calls `validate()` method internally. Therefore users of that object don't need to revalidate the object. The `GroupKey#validate` is now a private method.

### Future improvements

Is there a good reason to store the byte data of a `GroupKey` in two separate fields, which contain the same information in two different formats? 

Do we need `GroupKeyish` type?