refactor(client): `EncryptionUtil` names (#757)
Rename some methods in `EncryptionUtil` to be more specific about the encryption method:
- `encrypt` and `decrypt` -> `encryptWithAES` and `decryptWithAES`
- `encryptWithRSAPublicKey`, `decryptWithRSAPrivateKey` and `validateRSAPublicKey` now have `RSA` word in the names

The `encryptWithAES` and `decryptWithAES` methods are now public. The methods can be seen to be part of public API of that class, similarly to RSA-related methods. We also test the functionality of these method separately in unit tests.

Modified `encryptWithAES` and `decryptWithAES` method to get CipherKey as a parameter instead of GroupKey. This way there is a clear separation of abstraction level in the methods of EncryptionUtil: some methods operate on low-level encryption logic (buffers, cipher keys etc.), and other method operate on high-level classes like `StreamMessage` and `GroupKey`.

### Other refactoring

Renamed `RsaKeyPair` to `RSAKeyPair` so that we have consistent naming in RSA-related classes/methods.