test(client): Fake network node (#789)
Refactor the fake test environment so that we use the real `BrubeckNode` in the tests. Earlier the fake environment replaced `BrubeckNode` with a custom `FakeBrubeckNode`. Now we fake only the `NetworkNode`.

The client creates a network node instance by calling `NetworkNodeFactory#createNetworkNode`. In production environment, that just calls `createNetworkNode` from `streamr-network` package. In the fake environment we replace the factory with another factory which creates `FakeNetworkNode` instances.

### Other changes

- separated initialization and startup of a fake node (the node joins the network only if `start()` is called)
- fixed a bug in `FakeStorageNode` (it stored duplicate messages from multi-partitioned streams)
- removed obsolete await calls in `BrubeckNode` and `FakeStorageNode`
- renamed `ActiveNodes` to `FakeNetwork`