test(client): Convert end-to-end tests to integration tests (#795)
Converted most of the end-to-end tests to integration tests. As integration tests use the fake environment, the test are faster to run an easier to debug. Sometimes test setup is also simpler in the fake environment, and therefore tests can be more straightforward.

There is one simple end-to-end to test the publishing and subscribing between nodes (`publish-subscribe.test.ts`). Other end-to-end tests:
- integrate to the chain or The Graph (in docker dev environment), e.g. registries and stream search
- have some low level interaction with `NetworkNode` interaction, e.g. proxy publishing

The sharding of end-to-end test runs has been removed as the combined workflow is not significantly slower than other GitHub workflows (usually it takes approximately same time to run client's end-to-end and broker's end-to-end tests).

### Other changes

- Enhanced encryption tests by creating new test cases which reflect concrete use scenarios quite straightforwardly
- Removed `Validation.test.ts` and moved the test logic to a unit test (most logic was already covered by other tests)
- Moved some `TrackerRegistry` tests logic to protocol package, and removed the remaining test (it tests primarily the environment values)
- Added `parallel-publish.test.ts`
- Cleaned and simplified up some test code (e.g. removed unnecessary `connect()` calls)
- Small enhancements to `FakeStreamRegistry` and `FakeHttpUtils`

### Future improvements

- Reduce test duplication (same logic is sometimes covered by multiple tests)
- Combine or restructure some tests: e.g. `Resends.test.ts` and `Resends2.test.ts`
- Divide very generic test files, e.g. `StreamrClient.test.ts`
- Enable/remove skipped tests
- Unify naming of the test files